+++ tests/Ormolu/Fixity/PrinterSpec.hs
@@ -1,10 +1,15 @@
 {-# LANGUAGE RecordWildCards #-}
+{-# OPTIONS_GHC -fno-warn-orphans #-}
 
 module Ormolu.Fixity.PrinterSpec (spec) where
 
-import Data.Char qualified as Char
+import Data.List (intercalate)
+import Data.List.NonEmpty qualified as NE
 import Data.Map.Strict qualified as Map
 import Data.Text qualified as T
+import Distribution.ModuleName (ModuleName)
+import Distribution.ModuleName qualified as ModuleName
+import Distribution.Types.PackageName (PackageName, mkPackageName)
 import Ormolu.Fixity
 import Ormolu.Fixity.Parser
 import Ormolu.Fixity.Printer
@@ -12,26 +17,19 @@ import Test.Hspec
 import Test.Hspec.Megaparsec
 import Test.QuickCheck
 
-newtype FixityMapWrapper = FixityMapWrapper FixityOverrides
-  deriving (Show)
-
-instance Arbitrary FixityMapWrapper where
+instance Arbitrary FixityOverrides where
   arbitrary =
-    FixityMapWrapper . FixityOverrides . Map.fromList
+    FixityOverrides . Map.fromList
       <$> listOf ((,) <$> genOperator <*> genFixityInfo)
     where
-      scaleDown = scale (`div` 4)
       genOperator =
         OpName . T.pack <$> oneof [genNormalOperator, genIdentifier]
       genNormalOperator =
         listOf1 (scaleDown arbitrary `suchThat` isOperatorConstituent)
-      isOperatorConstituent x =
-        (Char.isSymbol x || Char.isPunctuation x) && x `notElem` ",`()"
       genIdentifier = do
-        x <- arbitrary `suchThat` Char.isLetter
+        x <- arbitrary `suchThat` isIdentifierFirstChar
         xs <- listOf1 (scaleDown arbitrary `suchThat` isIdentifierConstituent)
         return (x : xs)
-      isIdentifierConstituent x = Char.isAlphaNum x || x == '_' || x == '\''
       genFixityInfo = do
         fiDirection <-
           elements
@@ -42,9 +40,36 @@ instance Arbitrary FixityMapWrapper where
         fiPrecedence <- chooseInt (0, 9)
         return FixityInfo {..}
 
+instance Arbitrary ModuleReexports where
+  arbitrary = ModuleReexports . Map.fromListWith combine <$> listOf genReexport
+    where
+      combine x y = NE.sort (x <> y)
+      genReexport = do
+        exportingModule <- arbitrary
+        exports <- NE.sort . NE.fromList . getNonEmpty <$> scaleDown arbitrary
+        return (exportingModule, exports)
+
+instance Arbitrary PackageName where
+  arbitrary =
+    mkPackageName
+      <$> listOf1 (scaleDown arbitrary `suchThat` isPackageNameConstituent)
+
+instance Arbitrary ModuleName where
+  arbitrary =
+    ModuleName.fromString . intercalate "." <$> scaleDown (listOf1 genSegment)
+    where
+      genSegment = do
+        x <- arbitrary `suchThat` isModuleSegmentFirstChar
+        xs <- listOf (arbitrary `suchThat` isModuleSegmentConstituent)
+        return (x : xs)
+
+scaleDown :: Gen a -> Gen a
+scaleDown = scale (`div` 4)
+
 spec :: Spec
 spec = do
   describe "parseFixityOverrides & printFixityOverrides" $
     it "arbitrary fixity maps are printed and parsed back correctly" $
-      property $ \(FixityMapWrapper fixityMap) ->
-        parseFixityOverrides "" (printFixityOverrides fixityMap) `shouldParse` fixityMap
+      property $ \fixityOverrides moduleReexports ->
+        parseDotOrmolu "" (printDotOrmolu fixityOverrides moduleReexports)
+          `shouldParse` (fixityOverrides, moduleReexports)